+++ src/neptune_fetcher/read_only_project.py
@@ -19,6 +19,7 @@ __all__ = [
 
 import os
 import re
+from enum import Enum
 from typing import (
     TYPE_CHECKING,
     Dict,
@@ -56,8 +57,12 @@ from neptune.internal.id_formats import (
 )
 from neptune.management.internal.utils import normalize_project_name
 from neptune.objects.utils import prepare_nql_query
-from neptune.table import Table
+from neptune.table import (
+    Table,
+    TableEntry,
+)
 from neptune.typing import ProgressBarType
+from typing_extensions import Literal
 
 from neptune_fetcher.read_only_run import (
     ReadOnlyRun,
@@ -68,7 +73,7 @@ if TYPE_CHECKING:
     from pandas import DataFrame
 
 
-MAX_COLUMNS_ALLOWED = 10_000
+MAX_COLUMNS_ALLOWED = 100
 MAX_REGEXABLE_RUNS = 100
 
 
@@ -121,30 +126,22 @@ class ReadOnlyProject:
                 print(run)
             ```
         """
-        step_size = int(os.getenv(NEPTUNE_FETCH_TABLE_STEP_SIZE, "1000"))
+        yield from list_objects_from_project(self._backend, self._project_id, object_type="run")
 
-        query = NQLQueryAggregate(
-            items=[query_for_not_trashed(), query_for_runs_not_experiments()],
-            aggregator=NQLAggregator.AND,
-        )
+    def list_experiments(self) -> Generator[Dict[str, Optional[str]], None, None]:
+        """Lists all experiments of a project.
 
-        leaderboard_entries = self._backend.search_leaderboard_entries(
-            project_id=self._project_id,
-            types=[ContainerType.RUN],
-            query=query,
-            sort_by="sys/id",
-            step_size=step_size,
-            columns=["sys/id", "sys/custom_run_id"],
-            use_proto=True,
-        )
+        Returns a generator of dictionaries with experiment identifiers and names:
+        `{"sys/id": ..., "sys/custom_experiment_id": ..., "sys/name": ...}`.
 
-        for row in Table(
-            backend=self._backend, container_type=ContainerType.RUN, entries=leaderboard_entries
-        ).to_rows():
-            yield {
-                "sys/id": get_attribute_value_from_entry(entry=row, name="sys/id"),
-                "sys/custom_run_id": get_attribute_value_from_entry(entry=row, name="sys/custom_run_id"),
-            }
+        Example:
+            ```
+            project = ReadOnlyProject("workspace/project", api_token="...")
+            for experiment in project.list_experiments():
+                print(experiment)
+            ```
+        """
+        yield from list_objects_from_project(self._backend, self._project_id, object_type="experiment")
 
     def fetch_read_only_runs(
         self,
@@ -177,6 +174,20 @@ class ReadOnlyProject:
             ```
         """
         return self.fetch_runs_df(columns=["sys/id", "sys/custom_run_id"])
+
+    def fetch_experiments(self) -> "DataFrame":
+        """Fetches a table containing identifiers and names of experiments in the project.
+
+        Returns `pandas.DataFrame` with three columns (`sys/id`, `sys/custom_run_id` and `sys/name`)
+        and one row for each experiment.
+
+        Example:
+            ```
+            project = ReadOnlyProject("workspace/project", api_token="...")
+            df = project.fetch_experiments()
+            ```
+        """
+        return self.fetch_experiments_df(columns=["sys/id", "sys/custom_run_id", "sys/name"])
 
     def fetch_runs_df(
         self,
@@ -233,6 +244,116 @@ class ReadOnlyProject:
             specific_runs_df = my_project.fetch_runs_df(with_ids=specific_run_ids)
             ```
         """
+        return self._fetch_project_objects_df(
+            columns=columns,
+            columns_regex=columns_regex,
+            custom_id_regex=custom_id_regex,
+            with_ids=with_ids,
+            custom_ids=custom_ids,
+            states=states,
+            owners=owners,
+            tags=tags,
+            trashed=trashed,
+            limit=limit,
+            sort_by=sort_by,
+            ascending=ascending,
+            progress_bar=progress_bar,
+        )
+
+    def fetch_experiments_df(
+        self,
+        columns: Optional[Iterable[str]] = None,
+        columns_regex: Optional[str] = None,
+        names_regex: Optional[str] = None,
+        custom_id_regex: Optional[str] = None,
+        with_ids: Optional[Iterable[str]] = None,
+        custom_ids: Optional[Iterable[str]] = None,
+        states: Optional[Iterable[str]] = None,
+        owners: Optional[Iterable[str]] = None,
+        tags: Optional[Iterable[str]] = None,
+        trashed: Optional[bool] = False,
+        limit: Optional[int] = None,
+        sort_by: str = "sys/creation_time",
+        ascending: bool = False,
+        progress_bar: Union[bool, Optional[ProgressBarType]] = None,
+    ) -> "DataFrame":
+        """Fetches the experiments' metadata and returns them as a pandas DataFrame.
+
+        Args:
+            columns: None or a list of column names to include in the result.
+                Defaults to None, which includes all available columns up to 10k.
+            columns_regex: A regex pattern to filter columns by name.
+                Use this parameter to include columns in addition to the ones specified by the `columns` parameter.
+            names_regex: A regex pattern to filter the experiments by name.
+                When applied, it needs to limit the number of experiments to 100 or fewer.
+            custom_id_regex: A regex pattern to filter the experiments by custom ID.
+                When applied, it needs to limit the number of experiments to 100 or fewer.
+            with_ids: A list of experiment IDs to filter the results.
+            custom_ids: A list of custom experiment IDs to filter the results.
+            states: A list of experiment states to filter the results.
+            owners: A list of owner names to filter the results.
+            tags: A list of tags to filter the results.
+            trashed: Whether to return trashed experiments as the result.
+                If True: return only trashed experiments.
+                If False (default): return only non-trashed experiments.
+                If None: return all experiments.
+            limit: How many entries to return at most. If `None`, all entries are returned.
+            sort_by: Name of the field to sort the results by.
+                The field must represent a simple type (string, float, datetime, integer, or Boolean).
+            ascending: Whether to sort the entries in ascending order of the sorting column values.
+            progress_bar: Set to `False` to disable the download progress bar,
+                or pass a `ProgressBarCallback` class to use your own progress bar callback.
+
+        Returns:
+            DataFrame: A pandas DataFrame containing information about the fetched experiments.
+
+        Example:
+            ```
+            # Fetch all experiments with specific columns
+            columns_to_fetch = ["sys/modification_time", "training/lr"]
+            experiments_df = my_project.fetch_experiments_df(columns=columns_to_fetch, states=["active"])
+
+            # Fetch experiments by specific IDs
+            specific_experiment_ids = ["RUN-123", "RUN-456"]
+            specific_experiments_df = my_project.fetch_experiments_df(with_ids=specific_experiments_ids)
+            ```
+        """
+        return self._fetch_project_objects_df(
+            columns=columns,
+            columns_regex=columns_regex,
+            names_regex=names_regex,
+            custom_id_regex=custom_id_regex,
+            with_ids=with_ids,
+            custom_ids=custom_ids,
+            states=states,
+            owners=owners,
+            tags=tags,
+            trashed=trashed,
+            limit=limit,
+            sort_by=sort_by,
+            ascending=ascending,
+            progress_bar=progress_bar,
+            object_type="experiment",
+        )
+
+    def _fetch_project_objects_df(
+        self,
+        columns: Optional[Iterable[str]] = None,
+        columns_regex: Optional[str] = None,
+        names_regex: Optional[str] = None,
+        custom_id_regex: Optional[str] = None,
+        with_ids: Optional[Iterable[str]] = None,
+        custom_ids: Optional[Iterable[str]] = None,
+        states: Optional[Iterable[str]] = None,
+        owners: Optional[Iterable[str]] = None,
+        tags: Optional[Iterable[str]] = None,
+        trashed: Optional[bool] = False,
+        limit: Optional[int] = None,
+        sort_by: str = "sys/creation_time",
+        ascending: bool = False,
+        progress_bar: Union[bool, Optional[ProgressBarType]] = None,
+        object_type: Literal["run", "experiment"] = "run",
+    ) -> "DataFrame":
         step_size = int(os.getenv(NEPTUNE_FETCH_TABLE_STEP_SIZE, "200"))
 
         if columns is not None:
@@ -240,6 +361,8 @@ class ReadOnlyProject:
             columns = set(columns)
             columns.add("sys/id")
             columns.add("sys/custom_run_id")
+            if object_type == "experiment":
+                columns.add("sys/name")
 
             if columns_regex is not None:
                 columns = filter_columns_regex(
@@ -256,11 +379,20 @@ class ReadOnlyProject:
                     "Please limit the number of columns to 10 000 or fewer."
                 )
 
+        if names_regex is not None:
+            with_ids = filter_sys_name_regex(
+                names_regex=names_regex,
+                backend=self._backend,
+                project_qualified_name=self._project_qualified_name,
+                with_ids=with_ids,
+            )
+
         if custom_id_regex is not None:
             with_ids = filter_custom_id_regex(
                 custom_id_regex=custom_id_regex,
                 backend=self._backend,
                 project_qualified_name=self._project_qualified_name,
+                limit=limit,
                 with_ids=with_ids,
             )
 
@@ -271,6 +403,7 @@ class ReadOnlyProject:
             owners=owners,
             tags=tags,
             trashed=trashed,
+            is_run=object_type == "run",
         )
 
         leaderboard_entries = self._backend.search_leaderboard_entries(
@@ -324,11 +457,12 @@ def prepare_extended_nql_query(
             aggregator=NQLAggregator.AND,
         )
 
-    if is_run:
-        query = NQLQueryAggregate(
-            items=[query, query_for_runs_not_experiments()],
-            aggregator=NQLAggregator.AND,
-        )
+    items = [query, query_for_runs_not_experiments()] if is_run else [query, query_for_experiments_not_runs()]
+
+    query = NQLQueryAggregate(
+        items=items,
+        aggregator=NQLAggregator.AND,
+    )
 
     return query
 
@@ -349,6 +483,7 @@ def filter_columns_regex(
             experiment_ids_filter=with_ids,
         )
     )
+
     for field_definition in field_definitions:
         columns.add(field_definition.path)
 
@@ -359,6 +494,7 @@ def filter_custom_id_regex(
     custom_id_regex: str,
     backend: HostedNeptuneBackend,
     project_qualified_name: str,
+    limit: Optional[int],
     with_ids: Optional[List[str]] = None,
 ) -> List[str]:
     objects = paginate_over(
@@ -371,11 +507,17 @@ def filter_custom_id_regex(
     regex = re.compile(custom_id_regex)
     filtered_with_ids = []
 
+    should_continue = True
+
     for experiment in objects:
+        if not should_continue:
+            break
         for field in experiment.fields:
             if field.path == "sys/custom_run_id" and regex.match(field.value) is not None:
                 filtered_with_ids.append(experiment.object_key)
-
+                if len(filtered_with_ids) == limit:
+                    should_continue = False
+                    break
                 if len(filtered_with_ids) > MAX_REGEXABLE_RUNS:
                     raise ValueError(
                         "Too many runs matched the custom ID regex. "
@@ -404,3 +546,104 @@ def query_for_runs_not_experiments() -> NQLQuery:
         operator=NQLAttributeOperator.EQUALS,
         value="",
     )
+
+
+def query_for_experiments_not_runs() -> NQLQuery:
+    names = [(m.name, m.value) for m in NQLAttributeOperator]  # noqa
+
+    # handle the case when the client nql doesn't have the 'NOT_EQUALS' operator
+    if "NOT_EQUALS" not in NQLAttributeOperator.__members__:
+        names += [("NOT_EQUALS", "!=")]
+
+    operators = Enum("NQLAttributeOperator", names)
+
+    return NQLQueryAttribute(
+        name="sys/name",
+        type=NQLAttributeType.STRING,
+        operator=operators.NOT_EQUALS,  # noqa
+        value="",
+    )
+
+
+def list_objects_from_project(
+    backend: HostedNeptuneBackend,
+    project_id: UniqueId,
+    object_type: Literal["run", "experiment"],
+) -> List[Dict[str, Optional[str]]]:
+    step_size = int(os.getenv(NEPTUNE_FETCH_TABLE_STEP_SIZE, "1000"))
+    queries = [query_for_not_trashed()]
+    if object_type == "run":
+        queries.append(query_for_runs_not_experiments())
+    else:
+        queries.append(query_for_experiments_not_runs())
+
+    columns = ["sys/id", "sys/custom_run_id"]
+    if object_type == "experiment":
+        columns.append("sys/name")
+
+    query = NQLQueryAggregate(
+        items=queries,
+        aggregator=NQLAggregator.AND,
+    )
+
+    leaderboard_entries = backend.search_leaderboard_entries(
+        project_id=project_id,
+        types=[ContainerType.RUN],
+        query=query,
+        sort_by="sys/id",
+        step_size=step_size,
+        columns=columns,
+        use_proto=True,
+    )
+    return [
+        get_object_dictionary_from_row(row=row, object_type=object_type)
+        for row in Table(backend=backend, container_type=ContainerType.RUN, entries=leaderboard_entries).to_rows()
+    ]
+
+
+def get_object_dictionary_from_row(
+    row: TableEntry,
+    object_type: Literal["run", "experiment"],
+) -> Dict[str, Optional[str]]:
+    object_dict = {
+        "sys/id": get_attribute_value_from_entry(entry=row, name="sys/id"),
+        "sys/custom_run_id": get_attribute_value_from_entry(entry=row, name="sys/custom_run_id"),
+    }
+
+    if object_type == "experiment":
+        object_dict["sys/name"] = get_attribute_value_from_entry(entry=row, name="sys/name")
+
+    return object_dict
+
+
+def filter_sys_name_regex(
+    names_regex: str,
+    backend: HostedNeptuneBackend,
+    project_qualified_name: str,
+    with_ids: Optional[List[str]] = None,
+) -> List[str]:
+    objects = paginate_over(
+        getter=backend.query_fields_within_project,
+        extract_entries=lambda data: data.entries,
+        project_id=project_qualified_name,
+        field_names_filter=["sys/name"],
+        experiment_ids_filter=with_ids,
+    )
+    regex = re.compile(names_regex)
+    filtered_with_ids = []
+
+    for experiment in objects:
+        for field in experiment.fields:
+            if field.path == "sys/name" and regex.match(field.value) is not None:
+                filtered_with_ids.append(experiment.object_key)
+
+                if len(filtered_with_ids) > MAX_REGEXABLE_RUNS:
+                    raise ValueError(
+                        "Too many runs matched the names regex. "
+                        f"Please limit the number of runs to {MAX_REGEXABLE_RUNS} or fewer."
+                    )
+
+    if with_ids is None:
+        return filtered_with_ids
+    else:
+        return list(set(with_ids) & set(filtered_with_ids))