+++ tests/features/steps/project_steps.py
@@ -27,22 +27,37 @@ def step_impl(context):
     context.runs = list(context.project.list_runs())
 
 
+@when("we list experiments")
+def step_impl(context):
+    context.experiments = list(context.project.list_experiments())
+
+
 @then("runs list is not empty")
 def step_impl(context):
     assert context.runs != []
+
+
+@then("experiments list is not empty")
+def step_impl(context):
+    assert context.experiments != []
 
 
 @then("runs list contains the run details we created")
 def step_impl(context):
     assert sorted(context.runs, key=lambda kv: kv["sys/id"]) == [
         {
-            "sys/id": f"{context.project_key}-1",
-            "sys/name": "my-lovely-experiment",
-            "sys/custom_run_id": "fetcher-aa-1",
-        },
-        {
             "sys/id": f"{context.project_key}-2",
-            "sys/name": "",
             "sys/custom_run_id": "fetcher-bb-2",
+        },
+    ]
+
+
+@then("experiments list contains the experiment details we created")
+def step_impl(context):
+    assert sorted(context.experiments, key=lambda kv: kv["sys/id"]) == [
+        {
+            "sys/id": f"{context.project_key}-1",
+            "sys/custom_run_id": "fetcher-aa-1",
+            "sys/name": "my-lovely-experiment",
         },
     ]